requirements.txt
python-dotenv
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai
textstat
wordfreq